refactor: move Host into runtime-core
