+++ runtime/protocol/http/common/src/aws/smithy/kotlin/runtime/http/endpoints/functions/Functions.kt
@@ -5,8 +5,8 @@
 // This package implements common standard library functions used by endpoint resolvers.
 package aws.smithy.kotlin.runtime.http.endpoints.functions
 
+import aws.smithy.kotlin.runtime.net.*
 import aws.smithy.kotlin.runtime.util.InternalApi
-import aws.smithy.kotlin.runtime.util.net.*
 import aws.smithy.kotlin.runtime.util.text.ensureSuffix
 import aws.smithy.kotlin.runtime.util.text.urlEncodeComponent
