+++ runtime/runtime-core/common/src/aws/smithy/kotlin/runtime/net/Host.kt
@@ -2,7 +2,7 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
-package aws.smithy.kotlin.runtime.util.net
+package aws.smithy.kotlin.runtime.net
 
 import aws.smithy.kotlin.runtime.util.text.urlEncodeComponent
